werid bugs so it's a work in progress
